Add Fishtankp: Binding Native Elements with an Observer
Instead of adding all the fish components into a documentFragment
and rendering them in the "usual" DOM way, I am now using a wrapper
component that maintains that list.

I was going to actually drop a few vanilla input elements and bind
a few `addEventListeners`, re-rendering the fishTank with a new
array of fish that has been filtered down... but this is way more
convenient.

So yeah, while I'm on that topic, this commit also adds in a few
two-way data binded properties. In particular, I am making use of
the two-way data-bind via native elements.

https://www.polymer-project.org/1.0/docs/devguide/data-binding.html#two-way-native

I've also added a `filter` to the dom-repeate template, binding the
data-bound inputs to the computeFilter observer.

Example here:
https://www.polymer-project.org/1.0/docs/devguide/templates.html#dynamic-sort-and-filter-changes

BUT WAIT, theres more!
---

So I am binding a bunch of native elements to some properties that
end up updating a computed filter. It works great with just one,
but if you add a second property (in my case, isWinter), and change
that second input that is bound to the property... it does nothing.

Once I interact with the first input, however, the second input
begins to start working.  I have no earthly clue why things didn't
work out of the box as expected, but adding the notify property to
the properties I was binding seemed to immediately fix the problem.

More info: https://www.polymer-project.org/1.0/docs/migration.html#property-bindings

(maybe I'll file a PR against polymer)