tech(Security) Removing unused deps
This cl removes references to unused deps found in package.json. In a
previous version of this app I had used express and ejs to render the
pages. Seeing as that was unnecessary, I switched over to html, but
accidently left the `dependencies` in the package.json in tact.

I received a security vulnerability notice on one of the unused deps. In
the unlikely event someone forks this repo swaps things from the vanilla
html to using the ejs template system... I'd rather they have to
reinstall deps and get the newer versions without security problems.

Bonus points, I stop receiving these alerts. :)